Move JMS Message inti jms package.
