+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/service/event/jms/EventMessage.java
@@ -1,4 +1,6 @@
-package com.ubiqube.etsi.mano.service.event;
+package com.ubiqube.etsi.mano.service.event.jms;
+
+import com.ubiqube.etsi.mano.service.event.NotificationEvent;
 
 public class EventMessage {
 	NotificationEvent notificationEvent;
